Move star on github to sidebar
