+++ src/js/components/notifications.js
@@ -1,20 +1,12 @@
-import _ from 'underscore';
 import React from 'react';
 import ReactCSSTransitionGroup from 'react-addons-css-transition-group';
 import { connect } from 'react-redux';
-
-
-const shell = window.require('electron').shell;
 
 import AllRead from './all-read';
 import Oops from './oops';
 import Repository from './repository';
 
 export class NotificationsPage extends React.Component {
-  openBrowser() {
-    shell.openExternal('http://www.github.com/manosim/gitify');
-  }
-
   render() {
     const wrapperClass = 'container-fluid main-container notifications';
     const notificationsEmpty = this.props.notifications.isEmpty();
@@ -42,12 +34,6 @@ export class NotificationsPage extends React.Component {
             return <Repository repo={obj} repoName={repoSlug} key={repoSlug} />;
           })}
         </ReactCSSTransitionGroup>
-
-        {!_.isEmpty(groupedNotifications) && !this.props.hasStarred ? (
-          <div className="fork" onClick={this.openBrowser}>
-            <i className="fa fa-github" /> Star Gitify on GitHub
-          </div>
-        ) : null}
       </div>
     );
   }
@@ -55,7 +41,6 @@ export class NotificationsPage extends React.Component {
 
 function mapStateToProps(state) {
   return {
-    hasStarred: state.settings.hasStarred,
     failed: state.notifications.get('failed'),
     notifications: state.notifications.get('response')
   };
